+++ libHDiffPatch/HDiff/private_diff/limit_mem_diff/digest_matcher.h
@@ -89,7 +89,7 @@ private:
     void _search_cover(const hpatch_TStreamInput* newData,hpatch_StreamPos_t newOffset,
                        hpatch_TOutputCovers* out_covers,unsigned char* pmem,
                        void* dataLocker=0,void* newDataLocker=0);
-public: //private for muti-thread
+public: //private for multi-thread
     void _search_cover_thread(hpatch_TOutputCovers* out_covers,unsigned char* pmem,void* mt_data);
 };